pkp/pkp-lib#2187 general file stage required policy, and allow author access on revision files independent from uploader
