+++ classes/security/authorization/internal/SubmissionFileRequestedRevisionRequiredPolicy.inc.php
@@ -42,11 +42,6 @@ class SubmissionFileRequestedRevisionRequiredPolicy extends SubmissionFileBaseAc
 		$submissionFile = $this->getSubmissionFile($request);
 		if (!is_a($submissionFile, 'SubmissionFile')) return AUTHORIZATION_DENY;
 
-		// Make sure the file belongs to the submission in request.
-		$submission = $this->getAuthorizedContextObject(ASSOC_TYPE_SUBMISSION);
-		if (!is_a($submission, 'Submission')) return AUTHORIZATION_DENY;
-		if ($submission->getId() != $submissionFile->getSubmissionId()) return AUTHORIZATION_DENY;
-
 		// Make sure the file is part of a review round
 		// with a requested revision decision.
 		$reviewRound = $reviewRoundDao->getBySubmissionFileId($submissionFile->getFileId());
